feature Spaceship can move

+++ Projects/Domain/Domain/Model/SpaceshipCraftingRules.swift
@@ -9,7 +9,7 @@ import Foundation
 
 public struct SpaceshipCraftingRules {
     public var maxHealth: Int = 100
-    public var capacityMultiplier: Int = 10000
+    public var capacityMultiplier: Int64 = 10000
     public var universalSpaceTimeMultiplier: Int = 20
     public var durabilityTimeMultiplier: TimeInterval = 10
 }
